Redo Metharme tokenizing strategy
lol
